Refactored file structure for abstracting/unabstracting data, in preparation for class-based abstracting
