+++ yk_gmd_blender/yk_gmd/v2/converters/common/from_abstract.py
@@ -1,0 +1,353 @@
+from dataclasses import dataclass
+from typing import TypeVar, Tuple, List, Dict, Iterable, Callable, Set, Union
+
+from mathutils import Matrix
+
+from yk_gmd_blender.structurelib.base import FixedSizeArrayUnpacker
+from yk_gmd_blender.structurelib.primitives import c_uint16
+from yk_gmd_blender.yk_gmd.v2.abstract.gmd_attributes import GMDMaterial, GMDAttributeSet
+from yk_gmd_blender.yk_gmd.v2.abstract.gmd_mesh import GMDSkinnedMesh, GMDMesh
+from yk_gmd_blender.yk_gmd.v2.abstract.gmd_scene import depth_first_iterate, GMDScene
+from yk_gmd_blender.yk_gmd.v2.abstract.gmd_shader import GMDVertexBufferLayout
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_bone import GMDBone
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_node import GMDNode
+from yk_gmd_blender.yk_gmd.v2.abstract.nodes.gmd_object import GMDSkinnedObject, GMDUnskinnedObject
+from yk_gmd_blender.yk_gmd.v2.structure.common.checksum_str import ChecksumStrStruct
+from yk_gmd_blender.yk_gmd.v2.structure.common.node import NodeStackOp
+
+
+@dataclass(frozen=True)
+class RearrangedData:
+    ordered_nodes: List[Tuple[GMDNode, NodeStackOp]]
+    ordered_matrices: List[Matrix]
+    ordered_objects: List[Union[GMDSkinnedObject, GMDUnskinnedObject]]
+
+    root_node_indices: List[int]
+    node_id_to_node_index: Dict[int, int]
+    node_id_to_object_index: Dict[int, int]
+    object_id_to_matrix_index: Dict[int, int]
+
+    texture_names: List[ChecksumStrStruct]
+    texture_names_index: Dict[str, int]
+    shader_names: List[ChecksumStrStruct]
+    shader_names_index: Dict[str, int]
+    node_names: List[ChecksumStrStruct]
+    node_names_index: Dict[str, int]
+
+    # Tuple of (layout, layout_vertex_packing_flags, meshes)
+    vertex_layout_groups: List[Tuple[GMDVertexBufferLayout, int, List[GMDMesh]]]
+
+    ordered_meshes: List[GMDMesh]
+    mesh_id_to_index: Dict[int, int]
+    mesh_id_to_matrixlist: Dict[int, Tuple]
+    mesh_id_to_object_index: Dict[int, int]
+
+    ordered_attribute_sets: List[GMDAttributeSet]
+    attribute_set_id_to_index: Dict[int, int]
+    # List of [start, end_exclusive) ranges
+    attribute_set_id_to_mesh_index_range: Dict[int, Tuple[int, int]]
+
+    ordered_materials: List[GMDMaterial]
+    material_id_to_index: Dict[int, int]
+
+    # This is only for skinned meshes
+    mesh_matrixlist: List[Tuple]
+    # build with build_index_mapping(pool, key=tuple)
+    mesh_matrixlist_index: Dict[Tuple, int]
+    # packed_mesh_matrixlist: bytes
+    # packed_mesh_matrixlist_index: Dict[Tuple, int]
+    pass
+
+
+T = TypeVar('T')
+TKey = TypeVar('TKey')
+
+
+def build_index_mapping(pool: List[T], key: Callable[[T], TKey] = lambda x: x) -> Dict[TKey, int]:
+    return {
+        key(x): i
+        for i, x in enumerate(pool)
+    }
+
+
+def build_pools(strs: Iterable[str]) -> Tuple[List[ChecksumStrStruct], Dict[str, int]]:
+    # Given a set of strings, build the list of checksumstrs
+    # sort the list by checksum (?) - I think RGG do this
+    # return mapping of string -> index in list
+    pool = [ChecksumStrStruct.make_from_str(s) for s in strs]
+    pool.sort(key=lambda css: css.checksum)
+    return pool, build_index_mapping(pool, key=lambda css: css.text)
+
+
+def pack_mesh_matrix_strings(mesh_matrixlist: List[Tuple[int, ...]], pack_as_16bit: bool) -> Tuple[
+    bytes, Dict[Tuple, int]]:
+    matrixlist_bytearray = bytearray()
+    matrixlist_index = {}
+    for matrixlist in mesh_matrixlist:
+        matrixlist_index[matrixlist] = len(matrixlist_bytearray)
+
+        if pack_as_16bit:
+            matrixlist_bytearray += bytes([len(matrixlist)])
+            FixedSizeArrayUnpacker(c_uint16, len(matrixlist)).pack(big_endian=True, value=matrixlist,
+                                                                   append_to=matrixlist_bytearray)
+        else:
+            matrixlist_bytearray += bytes([len(matrixlist)] + list(matrixlist))
+
+    return bytes(matrixlist_bytearray), matrixlist_index
+
+
+def generate_vertex_layout_packing_flags(layout: GMDVertexBufferLayout) -> int:
+    return layout.packing_flags
+    # vertex_packing_flags = 0
+    #
+    # def pack_vector_type(vec_storage: Optional[VecStorage]) -> int:
+    #
+    #
+    # # pos_count = extract_bits(0, 3)
+    # # pos_precision = extract_bits(3, 1)
+    # # if pos_precision == 1:
+    # #     pos_storage = VecStorage.Vec3Half if pos_count == 3 else VecStorage.Vec4Half
+    # # else:
+    # #     pos_storage = VecStorage.Vec3Full if pos_count == 3 else VecStorage.Vec4Full
+    # if layout.pos_storage in [VecStorage.Vec3Full, VecStorage.Vec3Half]:
+    #     vertex_packing_flags |= 3
+    # else:
+    #     vertex_packing_flags |= 4
+    # if layout.pos_storage in [VecStorage.Vec3Half, VecStorage.Vec4Half]:
+    #     vertex_packing_flags |= (1 << 3)
+
+
+def arrange_data_for_export(scene: GMDScene) -> RearrangedData:
+    # Note - flags
+    # many bones flag is important, but so are the others - look into which ones are supposed to be there
+    # is relative-indexing set in a flag?
+
+    # ordered_nodes = []
+    # node_id_to_node_index = {}
+    # ordered_matrices = []
+    # node_id_to_matrix_index = {}
+    # root_node_indices = []
+    #
+    # skinned_objects = list(scene.skinned_objects.depth_first_iterate())
+    # unskinned_objects = list(scene.unskinned_objects.depth_first_iterate())
+
+    ordered_nodes = []
+    ordered_matrices = []
+    ordered_objects = []
+
+    root_node_indices = []
+    node_id_to_node_index = {}
+    node_id_to_object_index = {}
+    node_id_to_matrix_index = {}
+
+    texture_names = set()
+    shader_names = set()
+    node_names = set()
+
+    # Order the nodes
+    # Depth-first indexing
+    # Track touched nodes in set T(n)?
+    root_gmd_nodes = scene.overall_hierarchy.roots
+    for i, gmd_node in enumerate(depth_first_iterate(root_gmd_nodes)):
+        # stackop = none
+        # if has parent and all other children of your parent have been touched - stackop += pop
+        # the depth_first_iterate iterates through children in order
+        #   -> if we are the last child, all others must have been touched
+        want_pop = bool(gmd_node.parent) and gmd_node.parent.children[-1] is gmd_node
+        # if not leaf: stackop += push
+        want_push = bool(gmd_node.children)
+
+        stack_op = NodeStackOp.NoOp
+        if want_pop:
+            stack_op = NodeStackOp.Pop
+            if want_push:
+                stack_op = NodeStackOp.PopPush
+        elif want_push:
+            stack_op = NodeStackOp.Push
+
+        # emit (node, stackop)
+        ordered_nodes.append((gmd_node, stack_op))
+        node_id_to_node_index[id(gmd_node)] = i
+
+        # if node is instance of GMDObject (skinned or unskinned) add to ordered_objects
+        if isinstance(gmd_node, (GMDSkinnedObject, GMDUnskinnedObject)):
+            node_id_to_object_index[id(gmd_node)] = len(ordered_objects)
+            ordered_objects.append(gmd_node)
+
+        # if node is bone or unskinned, emit a matrix
+        if isinstance(gmd_node, (GMDBone, GMDUnskinnedObject)):
+            node_id_to_matrix_index[id(gmd_node)] = len(ordered_matrices)
+            ordered_matrices.append(gmd_node.matrix)
+
+        # if node has no parent, add index to roots
+        if not gmd_node.parent:
+            root_node_indices.append(i)
+
+        # Add name to node names
+        node_names.add(gmd_node.name)
+
+    # make sure to maintain original object order for scene
+    # involves making sure the DFA happens with objects in order
+
+    # collect meshes
+    meshes: List[GMDMesh] = [
+        mesh
+        for obj in ordered_objects
+        for mesh in obj.mesh_list
+    ]
+    for mesh in meshes:
+        shader_names.add(mesh.attribute_set.shader.name)
+
+        if mesh.attribute_set.texture_diffuse:
+            texture_names.add(mesh.attribute_set.texture_diffuse)
+        if mesh.attribute_set.texture_refl:
+            texture_names.add(mesh.attribute_set.texture_refl)
+        if mesh.attribute_set.texture_multi:
+            texture_names.add(mesh.attribute_set.texture_multi)
+        if mesh.attribute_set.texture_unk1:
+            texture_names.add(mesh.attribute_set.texture_unk1)
+        if mesh.attribute_set.texture_rs:
+            texture_names.add(mesh.attribute_set.texture_rs)
+        if mesh.attribute_set.texture_normal:
+            texture_names.add(mesh.attribute_set.texture_normal)
+        if mesh.attribute_set.texture_rt:
+            texture_names.add(mesh.attribute_set.texture_rt)
+        if mesh.attribute_set.texture_rd:
+            texture_names.add(mesh.attribute_set.texture_rd)
+
+    # build texture, shader, node name pools
+    texture_names, texture_names_index = build_pools(texture_names)
+    shader_names, shader_names_index = build_pools(shader_names)
+    node_names, node_names_index = build_pools(node_names)
+
+    # ordering meshes:
+    # build list of vertex buffer layouts to use
+    known_vertex_layouts_set: Set[GMDVertexBufferLayout] = {
+        mesh.vertices_data.layout
+        for mesh in meshes
+    }
+    # sort by descending flags int value (?)
+    known_vertex_layouts_and_flags = [(l, generate_vertex_layout_packing_flags(l)) for l in known_vertex_layouts_set]
+    known_vertex_layouts_and_flags.sort(key=lambda l_with_flags: l_with_flags[1], reverse=True)
+    vertex_layout_groups = []
+    for layout, flag in known_vertex_layouts_and_flags:
+        meshes_for_buffer = [m for m in meshes if m.attribute_set.shader.vertex_buffer_layout == layout]
+        # sort meshes by id(material) - just to group the common materials together
+        meshes_for_buffer.sort(key=lambda m: id(m.attribute_set))
+        # emit buffer_layout, meshes_for_buffer
+        vertex_layout_groups.append((layout, flag, meshes_for_buffer))
+
+    ordered_meshes = sum([ms for _, _, ms in vertex_layout_groups], [])
+    mesh_id_to_index = build_index_mapping(ordered_meshes, key=id)
+
+    mesh_id_to_object_index = {}
+    # These are only for skinned meshes
+    mesh_id_to_matrixlist = {}
+    mesh_matrixlist_set = set()
+    for object_idx, object in enumerate(ordered_objects):
+        for mesh in object.mesh_list:
+            if id(mesh) in mesh_id_to_object_index:
+                raise Exception(
+                    f"Mesh is mapped to two objects {object.name} and {ordered_objects[mesh_id_to_object_index[id(mesh)]].name}")
+            mesh_id_to_object_index[id(mesh)] = object_idx
+
+            if isinstance(object, GMDSkinnedObject):
+                if not isinstance(mesh, GMDSkinnedMesh):
+                    raise Exception(f"SkinnedObject {object.name} has unskinned mesh")
+                matrixlist = tuple([node_id_to_matrix_index[id(bone)] for bone in mesh.relevant_bones])
+                mesh_id_to_matrixlist[id(mesh)] = matrixlist
+                mesh_matrixlist_set.add(matrixlist)
+
+    mesh_matrixlist = list(mesh_matrixlist_set)
+    mesh_matrixlist_index = build_index_mapping(mesh_matrixlist)
+    # mesh_id_to_matrix_string_index = {
+    #     mesh_id:mesh_matrixlist.index(matrixlist)
+    #     for mesh_id, matrixlist in mesh_id_to_matrixlist.items()
+    # }
+
+    if set(mesh_id_to_index.keys()) != set(mesh_id_to_object_index.keys()):
+        raise Exception("Somehow the mapping of mesh -> mesh index maps different meshes than the mesh -> object index")
+
+    # Order the attribute sets
+    attribute_set_id_to_mesh_index_range = {}
+    ordered_attribute_sets = []
+    attr_index_start = -1
+    for i, m in enumerate(ordered_meshes):
+        if not ordered_attribute_sets:
+            ordered_attribute_sets.append(m.attribute_set)
+            attr_index_start = i
+        elif m.attribute_set != ordered_attribute_sets[-1]:
+            curr_attribute_set = ordered_attribute_sets[-1]
+            attr_index_end = i
+            attribute_set_id_to_mesh_index_range[id(curr_attribute_set)] = (attr_index_start, attr_index_end)
+            attr_index_start = i
+            ordered_attribute_sets.append(m.attribute_set)
+    attribute_set_id_to_mesh_index_range[id(ordered_attribute_sets[-1])] = (attr_index_start, len(ordered_meshes))
+
+    # make index mapping for ordered_materials
+    attribute_set_id_to_index = build_index_mapping(ordered_attribute_sets, key=id)
+
+    # Order the materials
+    material_ids = set()
+    ordered_materials = []
+    for attribute_set in ordered_attribute_sets:
+        if id(attribute_set.material) not in material_ids:
+            ordered_materials.append(attribute_set.material)
+            material_ids.add(id(attribute_set.material))
+    material_id_to_index = build_index_mapping(ordered_materials, key=id)
+
+    # TODO: Build drawlists?
+
+    # Build matrixlists
+    # mesh_matrix_index_list_set = set()
+    # mesh_id_to_matrixlist: Dict[int, List[int]] = {}
+    # for mesh in ordered_meshes:
+    #     if not isinstance(mesh, GMDSkinnedMesh):
+    #         continue
+    #
+    #     matrix_list = [node_id_to_node_index[id(bone)] for bone in mesh.relevant_bones]
+    #     mesh_matrix_index_list_set.add(tuple(matrix_list))
+    #     mesh_id_to_matrixlist[id(mesh)] = matrix_list
+    # mesh_matrixlist = [list(s) for s in mesh_matrix_index_list_set]
+    # mesh_matrixlist_index = build_index_mapping(mesh_matrixlist, key=tuple)
+
+    # now all arrangements should be made - next is to port into the respective file formats
+    # this is for tomorrow tho
+
+    return RearrangedData(
+        ordered_nodes=ordered_nodes,
+        ordered_matrices=ordered_matrices,
+        ordered_objects=ordered_objects,
+
+        root_node_indices=root_node_indices,
+        node_id_to_node_index=node_id_to_node_index,
+        node_id_to_object_index=node_id_to_object_index,
+        object_id_to_matrix_index=node_id_to_matrix_index,
+
+        texture_names=texture_names,
+        texture_names_index=texture_names_index,
+        shader_names=shader_names,
+        shader_names_index=shader_names_index,
+        node_names=node_names,
+        node_names_index=node_names_index,
+
+        # Tuple of (layout, layout_vertex_packing_flags, meshes)
+        vertex_layout_groups=vertex_layout_groups,
+
+        ordered_meshes=ordered_meshes,
+        mesh_id_to_index=mesh_id_to_index,
+        mesh_id_to_object_index=mesh_id_to_object_index,
+        # These are only for skinned meshes
+        mesh_id_to_matrixlist=mesh_id_to_matrixlist,
+        mesh_matrixlist=mesh_matrixlist,
+        mesh_matrixlist_index=mesh_matrixlist_index,
+
+        ordered_attribute_sets=ordered_attribute_sets,
+        attribute_set_id_to_index=attribute_set_id_to_index,
+        # List of [start, end_exclusive) ranges
+        attribute_set_id_to_mesh_index_range=attribute_set_id_to_mesh_index_range,
+
+        ordered_materials=ordered_materials,
+        material_id_to_index=material_id_to_index,
+
+    )
