+++ yk_gmd_blender/yk_gmd/v2/errors/error_classes.py
@@ -1,0 +1,6 @@
+class GMDImportExportError(Exception):
+    pass
+
+
+class InvalidGMDFormatError(GMDImportExportError):
+    pass
